pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aumos-content-provenance"
version = "0.1.0"
description = "AI content provenance with C2PA cryptographic tracking, invisible watermarking, training data lineage, license compliance, and court-admissible audit trails."
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.11"
authors = [{ name = "AumOS Enterprise" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "aumos-common>=0.1.0",
    "aumos-proto>=0.1.0",
    "fastapi>=0.110.0",
    "pydantic>=2.6.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.27.0",
    # C2PA and cryptographic provenance
    "cryptography>=42.0.0",
    "c2pa-python>=0.5.0",
    # Watermarking
    "Pillow>=10.2.0",
    "numpy>=1.26.0",
    "invisible-watermark>=0.2.0",
    # Storage and serialization
    "aiofiles>=23.2.1",
    "python-multipart>=0.0.9",
    # Kafka
    "aiokafka>=0.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "testcontainers[postgres,kafka]>=3.7.0",
    "httpx>=0.27.0",
    "factory-boy>=3.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/aumos_content_provenance"]

[tool.ruff]
line-length = 120
target-version = "py311"
src = ["src"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "ANN", "B", "A", "COM", "C4", "PT", "RUF"]
ignore = ["ANN101", "ANN102", "COM812"]

[tool.ruff.lint.isort]
known-first-party = ["aumos_content_provenance"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy", "sqlalchemy.ext.mypy.plugin"]
exclude = ["tests/"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=aumos_content_provenance --cov-report=term-missing --cov-fail-under=80"
